# topdown: Fix virtual cache to allow composite key terms