--- a/topdown/cache.go
+++ b/topdown/cache.go
@@ -6,6 +6,7 @@ package topdown
 
 import (
 	"github.com/open-policy-agent/opa/ast"
+	"github.com/open-policy-agent/opa/util"
 )
 
 type virtualCache struct {
@@ -14,7 +15,7 @@ type virtualCache struct {
 
 type virtualCacheElem struct {
 	value    *ast.Term
-	children map[ast.Value]*virtualCacheElem
+	children *util.HashMap
 }
 
 func newVirtualCache() *virtualCache {
@@ -34,12 +35,11 @@ func (c *virtualCache) Pop() {
 func (c *virtualCache) Get(ref ast.Ref) *ast.Term {
 	node := c.stack[len(c.stack)-1]
 	for i := 0; i < len(ref); i++ {
-		key := ref[i].Value
-		next := node.children[key]
-		if next == nil {
+		x, ok := node.children.Get(ref[i])
+		if !ok {
 			return nil
 		}
-		node = next
+		node = x.(*virtualCacheElem)
 	}
 	return node.value
 }
@@ -47,21 +47,28 @@ func (c *virtualCache) Get(ref ast.Ref) *ast.Term {
 func (c *virtualCache) Put(ref ast.Ref, value *ast.Term) {
 	node := c.stack[len(c.stack)-1]
 	for i := 0; i < len(ref); i++ {
-		key := ref[i].Value
-		next := node.children[key]
-		if next == nil {
-			next = newVirtualCacheElem()
-			node.children[key] = next
+		x, ok := node.children.Get(ref[i])
+		if ok {
+			node = x.(*virtualCacheElem)
+		} else {
+			next := newVirtualCacheElem()
+			node.children.Put(ref[i], next)
+			node = next
 		}
-		node = next
 	}
 	node.value = value
 }
 
 func newVirtualCacheElem() *virtualCacheElem {
-	return &virtualCacheElem{
-		children: map[ast.Value]*virtualCacheElem{},
-	}
+	return &virtualCacheElem{children: newVirtualCacheHashMap()}
+}
+
+func newVirtualCacheHashMap() *util.HashMap {
+	return util.NewHashMap(func(a, b util.T) bool {
+		return a.(*ast.Term).Equal(b.(*ast.Term))
+	}, func(x util.T) int {
+		return x.(*ast.Term).Hash()
+	})
 }
 
 // baseCache implements a trie structure to cache base documents read out of